long node ids

+++ hydra-data/src/main/java/com/addthis/hydra/data/tree/DataTree.java
@@ -33,7 +33,7 @@ public interface DataTree extends DataTreeNode {
 
     public void sync() throws IOException;
 
-    public int getDBCount();
+    public long getDBCount();
 
     public int getCacheSize();
